Modificações no chiSquare
Ordenação do Map e limitação de num de features

+++ src/com/datumbox/opensource/dataobjects/Document.java
@@ -31,7 +31,7 @@ public class Document {
     /**
      * List of token counts
      */
-    public Map<String, Integer> tokens;
+    public Map<String, Double> tokens;
     
     /**
      * The class of the document
